requirements.txt
openai
langchain
langchain-community
langchain-openai
langchain-core
streamlit
python-dotenv
pillow